feat: SearchActivity에서 PostNovelActivity로 이동하는 기능 추가

+++ app/src/main/java/com/teamwss/websoso/data/remote/response/SearchNovelsResponse.kt
@@ -12,7 +12,7 @@ data class SearchNovelsResponse(
     @Serializable
     data class Novel(
         @SerialName("novelId")
-        val novelId: Int,
+        val novelId: Long,
         @SerialName("novelImg")
         val novelImg: String,
         @SerialName("novelTitle")
